# Fix NodeSet Crash