## Description

    Fix rare crash in node set
    
    The bug was with faulty reasoning around calculating bucketB. In rare
    scenarios we get unlucky and the float64 mantissa cuts across at a
    bucket boundary. This will result in a "skipping" a bucket. This breaks
    invariants in the relationship between neighbouring buckets.


## Check List

Have you:

- Added unit tests? Yes

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- N/A

## Benchmark Results

There's about a 10% performance penalty for this change. Unfortunate, but this is a bug fix so not really negatiable. Can come back later and think of some alternate solutions that may be faster.

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-4                                              1.98ns ±20%    1.93ns ±19%     ~     (p=0.359 n=15+14)
LineEnvelope/1-4                                              1.28ns ±19%    1.24ns ± 8%     ~     (p=0.095 n=15+13)
LineEnvelope/2-4                                              1.62ns ±21%    1.53ns ±16%     ~     (p=0.072 n=15+15)
LineEnvelope/3-4                                              1.44ns ±19%    1.39ns ± 8%     ~     (p=0.064 n=15+14)
MarshalWKB/polygon/n=10-4                                      244ns ±14%     222ns ±16%   -8.81%  (p=0.002 n=12+15)
MarshalWKB/polygon/n=100-4                                     679ns ± 8%     628ns ±16%   -7.55%  (p=0.031 n=12+13)
MarshalWKB/polygon/n=1000-4                                   5.11µs ±62%    4.58µs ±30%     ~     (p=0.202 n=15+15)
MarshalWKB/polygon/n=10000-4                                  42.3µs ±73%    36.7µs ±70%     ~     (p=0.150 n=14+14)
UnmarshalWKB/polygon/n=10-4                                    382ns ±10%     360ns ±10%   -5.76%  (p=0.022 n=15+14)
UnmarshalWKB/polygon/n=100-4                                   844ns ±17%     804ns ±10%     ~     (p=0.156 n=13+12)
UnmarshalWKB/polygon/n=1000-4                                 4.94µs ±29%    4.37µs ±14%  -11.63%  (p=0.009 n=14+13)
UnmarshalWKB/polygon/n=10000-4                                47.4µs ±30%    39.5µs ±35%  -16.64%  (p=0.002 n=15+13)
IntersectsLineStringWithLineString/n=10-4                     1.82µs ±17%    1.72µs ±23%     ~     (p=0.194 n=14+14)
IntersectsLineStringWithLineString/n=100-4                    25.1µs ± 9%    23.7µs ± 6%   -5.86%  (p=0.001 n=13+13)
IntersectsLineStringWithLineString/n=1000-4                    247µs ±18%     231µs ± 6%   -6.72%  (p=0.015 n=14+12)
IntersectsLineStringWithLineString/n=10000-4                  3.60ms ±15%    3.31ms ±10%   -7.92%  (p=0.004 n=13+14)
IntersectsMultiPointWithMultiPoint/n=20-4                     1.61µs ±20%    1.55µs ±16%   -3.94%  (p=0.041 n=14+15)
IntersectsMultiPointWithMultiPoint/n=200-4                    16.9µs ±20%    16.2µs ±13%   -4.18%  (p=0.040 n=15+15)
IntersectsMultiPointWithMultiPoint/n=2000-4                    169µs ±18%     164µs ± 9%     ~     (p=0.065 n=15+13)
IntersectsMultiPointWithMultiPoint/n=20000-4                  1.80ms ±19%    1.74ms ±12%     ~     (p=0.057 n=14+15)
PolygonSingleRingValidation/n=10-4                            3.04µs ±26%    2.80µs ± 7%   -7.80%  (p=0.005 n=15+14)
PolygonSingleRingValidation/n=100-4                           39.3µs ±11%    36.0µs ± 4%   -8.53%  (p=0.000 n=13+12)
PolygonSingleRingValidation/n=1000-4                           474µs ±19%     462µs ± 6%     ~     (p=0.114 n=14+14)
PolygonSingleRingValidation/n=10000-4                         5.97ms ±10%    5.78ms ±10%     ~     (p=0.105 n=14+13)
PolygonMultipleRingsValidation/n=4-4                          8.07µs ± 8%    7.60µs ±11%   -5.89%  (p=0.003 n=12+13)
PolygonMultipleRingsValidation/n=36-4                         66.8µs ±17%    64.7µs ±20%     ~     (p=0.235 n=13+15)
PolygonMultipleRingsValidation/n=400-4                         897µs ±17%     838µs ± 7%   -6.49%  (p=0.002 n=13+14)
PolygonMultipleRingsValidation/n=4096-4                       10.3ms ±20%     9.9ms ± 6%     ~     (p=0.102 n=15+14)
PolygonZigZagRingsValidation/n=10-4                           13.1µs ±19%    12.5µs ±15%   -5.18%  (p=0.041 n=15+15)
PolygonZigZagRingsValidation/n=100-4                           152µs ±19%     148µs ±13%     ~     (p=0.233 n=15+15)
PolygonZigZagRingsValidation/n=1000-4                         1.81ms ± 8%    1.69ms ± 8%   -6.53%  (p=0.000 n=14+14)
PolygonZigZagRingsValidation/n=10000-4                        23.3ms ±25%    21.6ms ± 4%   -7.19%  (p=0.029 n=15+13)
PolygonAnnulusValidation/n=10-4                               4.37µs ±22%    4.20µs ±12%     ~     (p=0.089 n=15+15)
PolygonAnnulusValidation/n=100-4                              39.3µs ±14%    36.3µs ±13%   -7.74%  (p=0.002 n=15+15)
PolygonAnnulusValidation/n=1000-4                              658µs ±17%     633µs ± 8%     ~     (p=0.142 n=15+13)
PolygonAnnulusValidation/n=10000-4                            7.85ms ±17%    7.47ms ± 7%   -4.73%  (p=0.025 n=15+13)
MultipolygonValidation/n=1-4                                   465ns ±12%     435ns ±12%   -6.39%  (p=0.011 n=14+12)
MultipolygonValidation/n=4-4                                  1.05µs ±17%    1.03µs ±19%     ~     (p=0.334 n=14+14)
MultipolygonValidation/n=16-4                                 4.29µs ±17%    4.13µs ±15%     ~     (p=0.164 n=14+14)
MultipolygonValidation/n=64-4                                 19.3µs ±18%    18.5µs ±11%     ~     (p=0.093 n=15+14)
MultipolygonValidation/n=256-4                                 125µs ±24%     118µs ±10%     ~     (p=0.081 n=15+15)
MultipolygonValidation/n=1024-4                                581µs ±15%     556µs ±13%     ~     (p=0.077 n=15+14)
MultiPolygonTwoCircles/n=10-4                                 4.28µs ±16%    4.14µs ±18%     ~     (p=0.142 n=13+13)
MultiPolygonTwoCircles/n=100-4                                44.7µs ±15%    42.7µs ±11%     ~     (p=0.210 n=14+14)
MultiPolygonTwoCircles/n=1000-4                                484µs ±19%     461µs ± 9%     ~     (p=0.152 n=15+12)
MultiPolygonTwoCircles/n=10000-4                              6.81ms ±18%    6.60ms ±14%     ~     (p=0.250 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-4                      5.68µs ±28%    5.25µs ± 8%     ~     (p=0.114 n=14+14)
MultiPolygonMultipleTouchingPoints/n=10-4                     44.5µs ±15%    42.1µs ±16%   -5.42%  (p=0.037 n=14+15)
MultiPolygonMultipleTouchingPoints/n=100-4                     524µs ±17%     510µs ±12%     ~     (p=0.174 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000-4                   6.27ms ±17%    5.95ms ±13%     ~     (p=0.061 n=15+15)
WKTParsing/point-4                                            2.00µs ±20%    1.95µs ±19%     ~     (p=0.400 n=14+15)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           59.9µs ±16%    55.7µs ±13%   -7.02%  (p=0.029 n=15+14)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            58.4µs ±17%    56.5µs ± 9%     ~     (p=0.185 n=15+13)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           903µs ±17%     862µs ±11%     ~     (p=0.093 n=15+14)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            909µs ±26%     874µs ± 9%     ~     (p=0.331 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     6.53µs ±12%    6.32µs ±10%     ~     (p=0.158 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      6.59µs ±11%    6.25µs ±10%   -5.02%  (p=0.050 n=14+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    71.6µs ±16%    67.4µs ± 7%   -5.96%  (p=0.022 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     70.6µs ±16%    67.3µs ± 9%   -4.62%  (p=0.043 n=13+14)
MultiLineStringIsSimpleManyLineStrings/n=100-4                61.1µs ±16%    60.6µs ± 8%     ~     (p=0.494 n=14+12)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                655µs ±18%     625µs ±11%     ~     (p=0.072 n=15+13)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-4                          59.8µs ±22%    57.7µs ± 9%     ~     (p=0.399 n=15+12)
IntersectionWithoutValidation/n=100-4                          113µs ±19%     112µs ± 8%     ~     (p=0.477 n=15+14)
IntersectionWithoutValidation/n=1000-4                         470µs ±21%     467µs ±19%     ~     (p=0.806 n=15+15)
IntersectionWithoutValidation/n=10000-4                       4.20ms ±16%    4.18ms ±12%     ~     (p=0.943 n=15+12)
NoOp/n=10-4                                                   4.64µs ±19%    4.48µs ±15%     ~     (p=0.285 n=15+15)
NoOp/n=100-4                                                  14.7µs ±19%    14.3µs ±14%     ~     (p=0.205 n=15+15)
NoOp/n=1000-4                                                  108µs ±14%     102µs ± 9%   -5.43%  (p=0.014 n=14+13)
NoOp/n=10000-4                                                1.15ms ±25%    1.11ms ±39%     ~     (p=0.389 n=15+15)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-4                               2.50µs ±22%    2.43µs ±20%     ~     (p=0.249 n=15+13)
LineStringIsSimpleCircle/n=100-4                              35.5µs ±17%    36.1µs ±18%     ~     (p=0.949 n=14+15)
LineStringIsSimpleCircle/n=1000-4                              445µs ±20%     435µs ±10%     ~     (p=0.252 n=15+14)
LineStringIsSimpleCircle/n=10000-4                            5.68ms ±17%    5.51ms ± 8%     ~     (p=0.130 n=15+13)
LineStringIsSimpleZigZag/10-4                                 2.24µs ± 9%    2.18µs ± 8%     ~     (p=0.210 n=14+14)
LineStringIsSimpleZigZag/100-4                                35.6µs ±18%    34.9µs ±15%     ~     (p=0.400 n=14+15)
LineStringIsSimpleZigZag/1000-4                                427µs ±20%     424µs ±12%     ~     (p=0.477 n=15+14)
LineStringIsSimpleZigZag/10000-4                              5.73ms ±19%    5.69ms ± 8%     ~     (p=0.652 n=15+14)
SetOperation/n=4/Go_Intersection-4                            47.5µs ±19%    47.2µs ±14%     ~     (p=0.780 n=14+15)
SetOperation/n=4/Go_Difference-4                              49.2µs ±19%    48.4µs ± 4%     ~     (p=0.185 n=15+13)
SetOperation/n=4/Go_SymmetricDifference-4                     63.5µs ±17%    64.0µs ±10%     ~     (p=0.949 n=15+14)
SetOperation/n=4/Go_Union-4                                   50.9µs ±20%    51.7µs ±17%     ~     (p=0.935 n=15+15)
SetOperation/n=4/GEOS_Intersection-4                          58.1µs ±18%    56.5µs ±11%     ~     (p=0.161 n=15+15)
SetOperation/n=4/GEOS_Difference-4                            60.5µs ±21%    59.2µs ±14%     ~     (p=0.331 n=15+14)
SetOperation/n=4/GEOS_SymmetricDifference-4                   86.9µs ±21%    83.4µs ± 9%     ~     (p=0.108 n=15+13)
SetOperation/n=4/GEOS_Union-4                                 63.3µs ± 7%    59.3µs ±12%   -6.19%  (p=0.003 n=13+15)
SetOperation/n=8/Go_Intersection-4                            59.9µs ±18%    61.2µs ±17%     ~     (p=0.270 n=14+15)
SetOperation/n=8/Go_Difference-4                              59.5µs ±18%    61.1µs ±10%     ~     (p=0.720 n=14+13)
SetOperation/n=8/Go_SymmetricDifference-4                     78.8µs ±18%    79.8µs ±10%     ~     (p=0.683 n=15+14)
SetOperation/n=8/Go_Union-4                                   62.2µs ±20%    64.2µs ±12%     ~     (p=0.511 n=14+14)
SetOperation/n=8/GEOS_Intersection-4                          67.2µs ±20%    66.7µs ±13%     ~     (p=0.389 n=15+15)
SetOperation/n=8/GEOS_Difference-4                            68.4µs ±20%    67.9µs ± 9%     ~     (p=0.747 n=15+14)
SetOperation/n=8/GEOS_SymmetricDifference-4                    100µs ±18%      98µs ±10%     ~     (p=0.285 n=15+15)
SetOperation/n=8/GEOS_Union-4                                 68.0µs ±19%    67.5µs ±14%     ~     (p=0.624 n=15+15)
SetOperation/n=16/Go_Intersection-4                           80.4µs ±19%    83.5µs ±11%     ~     (p=0.377 n=15+14)
SetOperation/n=16/Go_Difference-4                             85.4µs ±20%    90.8µs ±18%     ~     (p=0.102 n=15+14)
SetOperation/n=16/Go_SymmetricDifference-4                     118µs ±16%     125µs ±10%     ~     (p=0.114 n=14+14)
SetOperation/n=16/Go_Union-4                                  89.0µs ±18%    95.2µs ±17%   +6.96%  (p=0.029 n=15+14)
SetOperation/n=16/GEOS_Intersection-4                         74.5µs ±19%    71.0µs ±10%     ~     (p=0.088 n=15+13)
SetOperation/n=16/GEOS_Difference-4                           80.8µs ±19%    81.4µs ±20%     ~     (p=1.000 n=15+15)
SetOperation/n=16/GEOS_SymmetricDifference-4                   128µs ±16%     123µs ± 5%   -3.72%  (p=0.032 n=15+12)
SetOperation/n=16/GEOS_Union-4                                83.6µs ±17%    81.3µs ± 9%     ~     (p=0.294 n=15+13)
SetOperation/n=32/Go_Intersection-4                            137µs ±19%     148µs ± 7%   +8.14%  (p=0.016 n=15+12)
SetOperation/n=32/Go_Difference-4                              144µs ±13%     152µs ±12%     ~     (p=0.114 n=14+14)
SetOperation/n=32/Go_SymmetricDifference-4                     199µs ±17%     208µs ±16%     ~     (p=0.270 n=15+14)
SetOperation/n=32/Go_Union-4                                   151µs ±15%     159µs ±21%     ~     (p=0.235 n=15+13)
SetOperation/n=32/GEOS_Intersection-4                         89.4µs ±19%    86.5µs ± 6%     ~     (p=0.183 n=15+12)
SetOperation/n=32/GEOS_Difference-4                           98.9µs ±17%    98.2µs ±10%     ~     (p=0.847 n=15+14)
SetOperation/n=32/GEOS_SymmetricDifference-4                   168µs ±19%     169µs ± 9%     ~     (p=0.870 n=15+15)
SetOperation/n=32/GEOS_Union-4                                 107µs ±21%     103µs ± 6%     ~     (p=0.330 n=15+11)
SetOperation/n=64/Go_Intersection-4                            229µs ±14%     249µs ±14%   +8.63%  (p=0.001 n=15+15)
SetOperation/n=64/Go_Difference-4                              247µs ±14%     265µs ±12%   +7.16%  (p=0.009 n=15+15)
SetOperation/n=64/Go_SymmetricDifference-4                     345µs ±14%     365µs ± 8%   +5.76%  (p=0.046 n=15+13)
SetOperation/n=64/Go_Union-4                                   259µs ±13%     282µs ±11%   +9.06%  (p=0.002 n=15+14)
SetOperation/n=64/GEOS_Intersection-4                          115µs ±15%     115µs ± 8%     ~     (p=0.648 n=15+12)
SetOperation/n=64/GEOS_Difference-4                            141µs ±16%     142µs ± 9%     ~     (p=0.914 n=15+14)
SetOperation/n=64/GEOS_SymmetricDifference-4                   260µs ±15%     256µs ±11%     ~     (p=0.174 n=15+15)
SetOperation/n=64/GEOS_Union-4                                 156µs ±19%     154µs ± 4%     ~     (p=0.867 n=15+12)
SetOperation/n=128/Go_Intersection-4                           426µs ±18%     456µs ±14%   +7.07%  (p=0.037 n=14+15)
SetOperation/n=128/Go_Difference-4                             461µs ±16%     487µs ± 8%   +5.66%  (p=0.015 n=15+13)
SetOperation/n=128/Go_SymmetricDifference-4                    615µs ±13%     654µs ±14%   +6.33%  (p=0.016 n=15+15)
SetOperation/n=128/Go_Union-4                                  465µs ±13%     509µs ± 6%   +9.34%  (p=0.001 n=15+13)
SetOperation/n=128/GEOS_Intersection-4                         177µs ±17%     177µs ±13%     ~     (p=0.949 n=15+14)
SetOperation/n=128/GEOS_Difference-4                           210µs ±17%     211µs ±14%     ~     (p=0.586 n=15+13)
SetOperation/n=128/GEOS_SymmetricDifference-4                  403µs ±12%     405µs ± 8%     ~     (p=1.000 n=15+15)
SetOperation/n=128/GEOS_Union-4                                232µs ±17%     234µs ± 8%     ~     (p=0.902 n=15+15)
SetOperation/n=256/Go_Intersection-4                           765µs ±15%     848µs ±12%  +10.87%  (p=0.001 n=15+13)
SetOperation/n=256/Go_Difference-4                             821µs ±16%     940µs ±16%  +14.48%  (p=0.000 n=13+15)
SetOperation/n=256/Go_SymmetricDifference-4                   1.19ms ±14%    1.29ms ±17%   +8.74%  (p=0.002 n=14+15)
SetOperation/n=256/Go_Union-4                                  874µs ±12%     956µs ±15%   +9.34%  (p=0.002 n=14+13)
SetOperation/n=256/GEOS_Intersection-4                         258µs ±16%     265µs ± 8%     ~     (p=0.316 n=15+13)
SetOperation/n=256/GEOS_Difference-4                           341µs ±17%     345µs ± 6%     ~     (p=0.751 n=15+13)
SetOperation/n=256/GEOS_SymmetricDifference-4                  740µs ±15%     774µs ± 7%     ~     (p=0.126 n=15+12)
SetOperation/n=256/GEOS_Union-4                                386µs ±15%     389µs ± 6%     ~     (p=0.892 n=15+13)
SetOperation/n=512/Go_Intersection-4                          1.64ms ±19%    1.72ms ±18%     ~     (p=0.106 n=15+15)
SetOperation/n=512/Go_Difference-4                            1.67ms ±15%    1.81ms ±16%   +8.80%  (p=0.001 n=14+15)
SetOperation/n=512/Go_SymmetricDifference-4                   2.31ms ± 8%    2.48ms ±11%   +7.42%  (p=0.003 n=14+15)
SetOperation/n=512/Go_Union-4                                 1.71ms ±11%    1.88ms ±13%   +9.79%  (p=0.001 n=15+14)
SetOperation/n=512/GEOS_Intersection-4                         459µs ±17%     473µs ±10%     ~     (p=0.331 n=15+14)
SetOperation/n=512/GEOS_Difference-4                           613µs ±19%     629µs ± 7%     ~     (p=0.270 n=15+14)
SetOperation/n=512/GEOS_SymmetricDifference-4                 1.35ms ±14%    1.42ms ±13%     ~     (p=0.093 n=14+15)
SetOperation/n=512/GEOS_Union-4                                680µs ±16%     714µs ± 7%     ~     (p=0.094 n=14+14)
SetOperation/n=1024/Go_Intersection-4                         2.94ms ±14%    3.37ms ±16%  +14.74%  (p=0.000 n=13+15)
SetOperation/n=1024/Go_Difference-4                           3.22ms ±14%    3.64ms ± 8%  +13.04%  (p=0.000 n=15+13)
SetOperation/n=1024/Go_SymmetricDifference-4                  4.55ms ±13%    5.02ms ±14%  +10.42%  (p=0.002 n=12+15)
SetOperation/n=1024/Go_Union-4                                3.36ms ±15%    3.71ms ± 9%  +10.68%  (p=0.000 n=14+14)
SetOperation/n=1024/GEOS_Intersection-4                        852µs ±14%     888µs ± 8%     ~     (p=0.098 n=15+13)
SetOperation/n=1024/GEOS_Difference-4                         1.21ms ±15%    1.29ms ±11%   +6.79%  (p=0.019 n=15+13)
SetOperation/n=1024/GEOS_SymmetricDifference-4                2.68ms ±14%    2.83ms ±11%     ~     (p=0.056 n=14+14)
SetOperation/n=1024/GEOS_Union-4                              1.40ms ±12%    1.47ms ± 9%   +4.33%  (p=0.044 n=14+14)
SetOperation/n=2048/Go_Intersection-4                         6.24ms ±17%    6.82ms ±15%   +9.35%  (p=0.003 n=15+14)
SetOperation/n=2048/Go_Difference-4                           6.71ms ±10%    7.35ms ±11%   +9.46%  (p=0.000 n=14+14)
SetOperation/n=2048/Go_SymmetricDifference-4                  9.41ms ± 8%   10.28ms ±19%   +9.27%  (p=0.001 n=13+14)
SetOperation/n=2048/Go_Union-4                                7.01ms ±14%    7.88ms ±17%  +12.36%  (p=0.000 n=15+15)
SetOperation/n=2048/GEOS_Intersection-4                       1.71ms ±16%    1.81ms ± 8%     ~     (p=0.051 n=15+14)
SetOperation/n=2048/GEOS_Difference-4                         2.21ms ±17%    2.27ms ±10%     ~     (p=0.310 n=15+14)
SetOperation/n=2048/GEOS_SymmetricDifference-4                5.22ms ±18%    5.41ms ±14%     ~     (p=0.148 n=15+15)
SetOperation/n=2048/GEOS_Union-4                              2.59ms ±16%    2.66ms ±19%     ~     (p=0.354 n=14+15)
SetOperation/n=4096/Go_Intersection-4                         13.2ms ±27%    14.8ms ±14%  +12.46%  (p=0.003 n=15+15)
SetOperation/n=4096/Go_Difference-4                           13.8ms ±11%    15.5ms ±10%  +12.63%  (p=0.000 n=15+15)
SetOperation/n=4096/Go_SymmetricDifference-4                  19.5ms ±16%    22.0ms ±14%  +12.67%  (p=0.000 n=14+14)
…